Extract session model and workflow controller
